seperated hooks for watchlist
